fix: coupon campaign add

+++ frontend/plugins/loyalty_ui/src/modules/loyalties/settings/coupon/graphql/mutations/CouponMutations.tsx
@@ -15,7 +15,7 @@ export const CREATE_COUPON_CAMPAIGN = gql`
     $codeRule: JSON
     $restrictions: JSON
     $redemptionLimitPerUser: Int
-    $buyScore: Float
+    $buyScore: Int
     $charSet: [String]
   ) {
     couponCampaignAdd(
